App Meta Prototype Version

+++ app/src/main/java/com/eae/busbarar/data/model/Tempreture.kt
@@ -7,5 +7,5 @@ import kotlinx.parcelize.Parcelize
 data class Temperature(
     val datetime : String?,
     val pred: String?,
-    val value: Int?
+    val value: Float?
 ) : Parcelable
